+++ source/benoit/benoit/application/handle_keys.rs
@@ -21,7 +21,7 @@
 	If not, see <https://www.gnu.org/licenses/>.
 */
 
-use crate::benoit::{Fractal, PRECISION};
+use crate::benoit::PRECISION;
 use crate::benoit::application::Application;
 
 extern crate rug;
@@ -32,20 +32,11 @@ use sdl2::keyboard::Scancode;
 
 impl Application {
 	pub fn handle_keys(&mut self, scan_code: Scancode) -> bool {
-		let print_values = |fractal: Fractal, center_real: &Float, center_imaginary: &Float, zoom: &Float, maximum_iteration_count: u32, julia_real: &Float, julia_imaginary: &Float| {
-			let base = format!("c = {center_real}{center_imaginary:+}i -- {zoom}x @ {maximum_iteration_count} iter.");
-
-			match fractal {
-				Fractal::Julia      => eprintln!("{base}:\n  c(M) = {julia_real}{julia_imaginary:+}i"),
-				Fractal::Mandelbrot => eprintln!("{base}"),
-			};
-		};
-
 		match scan_code {
 			Scancode::Escape => return true,
 			Scancode::C      => self.do_draw = true,
 			Scancode::X      => self.do_dump = true,
-			Scancode::Z      => print_values(self.fractal, &self.center_real, &self.center_imaginary, &self.zoom, self.maximum_iteration_count, &self.julia_real, &self.julia_imaginary),
+			Scancode::Z      => eprintln!("c = {}{:+}i -- {}x @ {} iter.", &self.center_real, &self.center_imaginary, &self.zoom, self.maximum_iteration_count),
 			_                => {},
 		}
 
@@ -79,18 +70,6 @@ impl Application {
 			Scancode::F => self.maximum_iteration_count * 0x2,
 			Scancode::R => self.maximum_iteration_count / 0x2,
 			_           => self.maximum_iteration_count,
-		};
-
-		match scan_code {
-			Scancode::Down  => self.julia_real += &translate_ammount,
-			Scancode::Up    => self.julia_real -= &translate_ammount,
-			_               => {},
-		};
-
-		match scan_code {
-			Scancode::Left  => self.julia_imaginary -= &translate_ammount,
-			Scancode::Right => self.julia_imaginary += &translate_ammount,
-			_               => {},
 		};
 
 		return false;