+++ source/benoit/benoit/application/render.rs
@@ -32,7 +32,7 @@ use std::time::Instant;
 use std::ptr::addr_of_mut;
 
 impl Application {
-	pub fn render(&self, buffer: &mut [u32], center_real: &Float, center_imaginary: &Float, zoom: &Float, maximum_iteration_count: u32, julia_real: &Float, julia_imaginary: &Float) {
+	pub fn render(&self, buffer: &mut [u32], center_real: &Float, center_imaginary: &Float, zoom: &Float, maximum_iteration_count: u32) {
 		eprint!("rendering...");
 
 		let mut threads = Vec::<JoinHandle<()>>::with_capacity(self.thread_count as usize);
@@ -68,10 +68,7 @@ impl Application {
 				let center_imaginary = center_imaginary.clone();
 				let zoom             = zoom.clone();
 
-				let julia_real      = julia_real.clone();
-				let julia_imaginary = julia_imaginary.clone();
-
-				threads.push(spawn(move || { render_row(buffer_slice, y, canvas_width, canvas_height, center_real, center_imaginary, zoom, maximum_iteration_count, julia_real, julia_imaginary) }));
+				threads.push(spawn(move || { render_row(buffer_slice, y, canvas_width, canvas_height, center_real, center_imaginary, zoom, maximum_iteration_count) }));
 
 				y += 0x1;
 			}
@@ -92,10 +89,7 @@ impl Application {
 				let center_imaginary = center_imaginary.clone();
 				let zoom             = zoom.clone();
 
-				let julia_real      = julia_real.clone();
-				let julia_imaginary = julia_imaginary.clone();
-
-				threads.push(spawn(move || { render_row(buffer_slice, y, canvas_width, canvas_height, center_real, center_imaginary, zoom, maximum_iteration_count, julia_real, julia_imaginary) }));
+				threads.push(spawn(move || { render_row(buffer_slice, y, canvas_width, canvas_height, center_real, center_imaginary, zoom, maximum_iteration_count) }));
 			}
 		}